test: move utf16beBuffer helper to test/util.ts
