+++ test/decodeFileContent.test.ts
@@ -1,17 +1,6 @@
 import { describe, test, expect } from 'vitest';
 import { decodeFileContent } from '../src/util';
-
-// UTF-16BE has no built-in Node Buffer encoding, so derive it by
-// byte-swapping the UTF-16LE encoding of the same string.
-function utf16beBuffer(contents: string): Buffer {
-  const le = Buffer.from(contents, 'utf16le');
-  const be = Buffer.alloc(le.length);
-  for (let i = 0; i < le.length; i += 2) {
-    be[i] = le[i + 1];
-    be[i + 1] = le[i];
-  }
-  return be;
-}
+import { utf16beBuffer } from './util';
 
 describe('with a UTF-16LE byte-order mark', () => {
   test('decodes as UTF-16LE and strips the BOM', () => {
